+++ addOns/scripts/src/main/java/org/zaproxy/zap/extension/scripts/automation/actions/EnableableScriptAction.java
@@ -41,7 +41,9 @@ public class EnableableScriptAction extends ScriptAction {
                     ScriptJobDialog.SCRIPT_INLINE_PARAM,
                     ScriptJobDialog.SCRIPT_IS_INLINE_PARAM,
                     ScriptJobDialog.SCRIPT_FILE_PARAM,
-                    ScriptJobDialog.SCRIPT_TARGET_PARAM);
+                    ScriptJobDialog.SCRIPT_TARGET_PARAM,
+                    ScriptJobDialog.SCRIPT_CONTEXT_PARAM,
+                    ScriptJobDialog.SCRIPT_USER_PARAM);
 
     public EnableableScriptAction(String name, boolean enabled, ScriptJobParameters parameters) {
         super(parameters);